[CHORE] 자잘한 count 수정

+++ src/main/java/app/nook/record/dto/RecordResponseDto.java
@@ -11,7 +11,7 @@ public class RecordResponseDto {
     }
 
     public record RecordCountDto(
-            int count
+            long count
     ) {
     }
